requirements.txt
# Django Core
Django==5.2.12
djangorestframework==3.14.0
django-environ==0.10.0
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

# Google OAuth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Database
psycopg2-binary==2.9.9

# PDF Generation
reportlab==4.0.7
openpyxl==3.1.5
pandas==2.1.3

# WhatsApp Integration (Opcional - requiere Windows Long Path support habilitado)
twilio==8.10.0

# Image Processing
Pillow>=10.0.0
qrcode[pil]==7.4.2

# Development (opcional)
django-extensions==3.2.3
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Production (opcional)
gunicorn==21.2.0
whitenoise==6.6.0

# API Documentation
drf-spectacular==0.27.2

# Async Task Queue
celery==5.3.6
redis==5.0.1
django-celery-results==2.5.1

# Performance Testing
locust==2.24.0
